[ADD] crm_claim_ref_smartbutton: Smart-button for referenced claims
Smart-button for referenced claims
==================================

This module adds an smart-button in the models that can be referenced in a
claim with the number of linked claims and a shortcut to them.

Usage
=====

Go to any record that can be referenced in a claim, and you will see in a
smart-button called _linked claims_ a number indicating the number of
claims that reference this record. Clicking on the button, it will navigate
to a list of these claims.

If you click _New_ on the claim list, the created claim will be automatically
linked to the referenced record.

Known issues / Roadmap
======================

* The check for seeing if the opened model is one of the possible referenced
  models is done on each view request, so this hits very little, but a bit,
  the overall performance of the system.