[ADD] crm_lead_substage: Substages for leads/opportunities
=================================
Substages for leads/opportunities
=================================

This module allows to set substages inside the same stage for fine-graining
the reasons or flows on the CRM pipelines.

Configuration
=============

Go to Sales > Configuration > Leads & Opportunities > Substages for defining
the substages you want and set the stages where this substages are going to be
available

Usage
=====

Go to Sales > Leads or Sales > Opportunities, and when you are in one stage,
the available substages are shown below. Click on one of them to set it. If you
change the stage, then the substage is reset and the new ones available for
this stage are shown.