Ordered update of fields
Without respecting the order, the user could get to a situation where
you are updating a field without lastname to set `firstname=False,
lastname="something"` and fail because Odoo started by `firstname` and
that lead to a temporary situation where no names are set and
`EmptyNamesError` gets raised.

With this patch, you force Odoo to write first the field with value,
whith avoids that situation and always works.