Better way to remove sale-related buttons in lead form view
Previously we inherited the view and removed the button box completely,
even for sales manager users and even the buttons which didn't require
access to the sales models.
This is a much better solution that just makes the changes introduced
by the sale_crm module restricted to users with a level of group_sale_salesman
and above; the result is that shared leads users can open the leads form
and without errors and still see the other buttons (e.g. 'Meetings')
while real sale users can still all the sale-related buttons.