[ADD][12.0] crm_opportunity_currency
This module allows to select a specific currency for an opportunity.
If the selected currency is different to the company currency,
an amount in the customer currency can be set and the expected revenue
of the opportunity will automatically be computed in the company currency.

In kanban view the amount and currency is displayed accordingly.