[IMP] crm_claim: remove unnecessary fields in views.
We didn't have activities back in v8 but the fields `action_next` and `date_deadline` in the models to be set from the views. In the later migration to v11 those view fields were adapted to the proper similar ones in activities. But that's useless, as activities have their own UI ready to be used from any model with the mixin.

TT38822